moved app js to components
